Move iterator code to a single module
For the Django 4 upgrade fix these objects get more dependency,
so its better to have them in a separate module. This also keeps all
iterator magic in a single place.
